+++ doc/source/changelog.rst
@@ -80,6 +80,7 @@ CMake
 * Added :cmake:function:`qi_generate_src`,  :cmake:function:`qi_generate_header`
 * Added :cmake:function:`qi_swig_wrap_java`
 * Added :cmake:function:`qi_install_python`
+* Added :cmake:function:`qi_stage_dir`, to find files from source directories
 * Added :cmake:function:`qi_create_perf_test`
 * :cmake:function:`qi_create_gtest` and :cmake:function:`qi_create_test` learned
   the ``SLOW`` keyword so that tests are not run by default.